Layoutlib: Read and close XML files as soon as possible. do not merge.
Because passing an InputStream to KXML does not close the
stream after the file has been parsed, the files are staying
locked on windows until the gc and finalizers are run.

This change preload the XML files and close their stream,
and then pass the content in a stream to the parser.

(cherry picked from commit 3b92caa0b1d15f4c80a4f77d8d0fbec8f24a1261)

Change-Id: I1b225454c166026c1aab6290c55ac8279e8862d5