Impose a height & width downward on challenge views.
This means that challenges may no longer wrap_content, even
vertically. But it's the only way to rationalize across
different challenges, some of which want to be a fixed size,
some match_parent, and some wrap_content.

This change also fixes some problems with the password view
as well as PIN view (usable again on sw360 devices like
Galaxy Nexus).

Bug: 7427586
Change-Id: I648e68c5159fef7cf594d4bfe8f6d4bdd5da6b6a
Proto-Id: I280b4875a9deee5dfb7409bdfa8ed0b96fbd0cff