Revert "NumberPicker should adjust min and max when displayed values are set." (a.k.a. Santa is back)
This reverted change was adjusting the min and max values for the NumberPicker
which is not desirable since it changes behavior and it will be possible for
an app that works on the current platform to crash on an older one. Also the
adjustment was not implemented correctly.

Updated the documentation to clarify the reltionship between the min value,
max value, and the displayed values array.

Bug:7518172

This reverts commit f84ebfa5b31926b47c84865d396d6030ac6810af

Change-Id: I109f1b1f54c1e609941243cabab9241871b6b12b