DO NOT MERGE. cherrypick from I3653f67ee321eac37cb8a6d228b1ef6d104ff0be
Add single crunch command to aapt. do not merge.

Previously the crunch command would work on a full res folder
and output a full res folder (with only the drawables). This
was only used in the SDK.

The incremental logic is moved to the SDK build system so we
change the crunch command (or rather add a new one) to only
crunch a single file.

(cherry picked from commit bdfa1e5b2b85282609977448b88642a3740aef9f)

Change-Id: Ifa0dceed93708658f950322f36c13b6ebe7ab0f3